feat: upload files and give public URLs (#7)
* feat: upload files and give public URLs

This is a composite of two changes:

* src/tools/buckets.ts: attach isPublic metadata to listed buckets
* src/tools/objects.ts: implement upload_file_and_get_url tool

This changes the list_buckets call to include an `isPublic` value
that is set to true if the bucket is public as defined by the ACL
grants including the URI http://acs.amazonaws.com/groups/global/AllUsers.

The `upload_file_and_get_url` tool uses this `isPublic` value to
determine if it should generate a presigned URL or not. If the bucket
is private, then it will generate a presigned URL, if not then it
generates a public one.

This enables the workflow of uploading a file to a bucket and then
getting a URL so you can share it with your friends.

Signed-off-by: Natalie Tersyn <natahul@yolo-swag.com>

* feat: allow setting AWS_REGION to point to any region

Signed-off-by: Natalie Tersyn <natahul@yolo-swag.com>

* fix(buckets/objects): remove bucket publicity checks

This was causing an issue with keys that were not registered as
the owner of an organization or keys that did not have the admin
bit set. This changes the upload and give URL tool to always return
a presigned URL.

Can't say I didn't try.

Signed-off-by: Natalie Tersyn <natahul@yolo-swag.com>

---------

Signed-off-by: Natalie Tersyn <natahul@yolo-swag.com>